-Fixed readwrite.h to work on LP64 systems. Thanks Amitai Schlair <schmonz@schmonz.com>

+++ readwrite.h
@@ -1,7 +1,8 @@
 #ifndef READWRITE_H
 #define READWRITE_H
 
-extern int read();
-extern int write();
+#include <sys/types.h>
+extern ssize_t read();
+extern ssize_t write();
 
 #endif
